refactor into views
